requirements.txt
# Install into repo .venv: ./scripts/bootstrap-python.sh — run Python via ./scripts/py (see AGENTS.md).
PyYAML>=6.0
